pyproject.toml
[project]
name = "fastapi-queue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.116.1",
    "pydantic-settings>=2.10.1",
    "supabase>=2.18.1",
    "uvicorn[standard]==0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.11",
]
